requirements.txt
streamlit>=1.29.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph>=0.0.10
transformers>=4.30.0
numpy>=1.24.0
uuid>=1.30
aiohttp>=3.8.0
asyncio>=3.4.3
textblob>=0.17.1